app/requirements.txt
google-generativeai
langchain>=0.1.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
feedparser>=6.0.10
psycopg2-binary>=2.9.9
tenacity>=8.2.3